embed constraint version files at compile time
